`bin/setup`: Join with `heroku apps:info`
Adding a local remote for the Heroku environments with `heroku join`
consistently fails. In the time since this line was first introduced,
the semantics of the CLI's heroku join call might have changed.

Since this line serves as a guard to ensure that the local machine has
proper access to the Heroku application, replace `heroku join` with
`heroku apps:info` to check for whether or not the command line has
access to the deployment's metadata.

Since we're interested in the exit code of `heroku apps:info` and not
the content, pipe the output to `/dev/null`.